requirements.txt
asgiref==3.7.2
Django==4.2
django-filter==23.5
django-phonenumber-field==7.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
phonenumberslite==8.13.30
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
